requirements.txt
dataclasses==0.6
Django==2.2
djangorestframework==3.10.0
djongo==1.2.33
pkg-resources==0.0.0
pymongo==3.8.0
pytz==2019.1
sqlparse==0.2.4
